[workaround] Change in the code to workaround the @opensums bug in the typechecker and avoid Magic.id.
Instead of brute cast the player, we apply on it a type constraint (of its
actual and current type), then open its sum type. This is a workaround against
the current bug of the typechecker until we really understand and fix the
issue.
Avoiding a Magic.id makes the code cleaner.